requirements.txt
matplotlib==3.6.3
numpy==1.23.5
pandas==1.5.2
Pillow==9.5.0
plotly==5.13.0
scikit_learn==1.2.2
seaborn==0.12.2
spacy==3.5.0
streamlit==1.15.2
streamlit_option_menu==0.3.2
